refactor: extract frontend account test to own folder
